CRIMAP-14 Prepare for cookies consent
This first PR implements the new cookies controller, helpers and the handling of the cookie settings.

The banner is not shown yet on the page, this is only the backend work.

In a follow-up PR we will enable the banner and its functionality on the frontend following the guidelines and design stated in the design system:
https://design-system.service.gov.uk/components/cookie-banner/