CRIMAP-14 Show cookie consent banner
Show the cookie consent banner on page load, and hook the accept/reject buttons to the controller.

Add a bit of javascript to allow to close the banner without reloading the page. With JS disabled, it will reload the page.

This follows the guidelines and design stated in the design system:
https://design-system.service.gov.uk/components/cookie-banner/

Also added the cookie settings to the cookies page, as per
https://design-system.service.gov.uk/patterns/cookies-page/

Content of the cookies page is best effort, guidelines, and based on what other services do, but may require a review of the copy.

Reorganised a bit some locales.