Tweaks to header colour handling and smoke test
A few tweaks that makes it more robust.

The CSS will in effect use the orange colour for any environment unless it is production (this is better than having to specify one by one the non-prod envs).

The phase tag text will also default to whatever is the name of the env, unless overridden by locales (as it is the case of `production` -> `alpha`).
This is also better than having to specify one by one all the env names, as usually the phase tag just shows the same (`staging` -> `staging`, `local` -> `local`).

Added a simple smoke test to ensure the CSS classes are applied correctly and the phase tag text is also correct.