Allow bypass with an under 18 (developer tools)
Previously there was only one "bypass DWP", which picks a random "mock" user (dob, nino, surname) and creates the record.

However there was a mock under18 so randomly it could happen this under18 user was picked for bypass, leaving the underlying record with inconsistent data (like having NINO, or the `passporting_benefit` flag set to true), which is incorrect for under 18s as we don't ask for NINO nor perform DWP check for those.

Removed this mock user as is no longer needed (no DWP is performed for under 18 anyways) and added a new button to explicitly trigger a bypass of an under 18, which will set everything correctly and even advance a bit more, up to the case type step.