Bump the schemas gem to 0.1.9
This version has some breaking changes, mainly for the time being we will only be able to submit single-class (or manually entered) offences, until the split of offences is finalised and the new list is incorporated to the service.

Fixed a few test, one of them for this reason (Robbery was C/B class, changed to a single-class offence).