CRIMAP-375 Cover additional passporting edge cases (#361)
This is an evolving piece of work so we might have to tweak some things later on.

But for now (bear in mind most of this PR are refactored tests), this will ensure proper handling of dependent attributes on the `client details` page, and the `NINO` page.

This covers a few scenarios that most likely will never happen in real life but it is easy to trigger these when doing test applications. In particular, the one where DWP was successful, and changing the NINO keeps saying it is successful (as it will take previous value and not re-trigger the check).