CRIMAP-373 Use new single-class offences list
And finalise some remaining work related to the IoJ passport based on offence, as this new list, in addition to be single-class, also introduce the new column `ioj_passport` (true / false).

Cleanup some old code related to multi-class (presenter, locales, tests, etc.)

Note, as a separate PR, I will add the new `passportable` attribute to the schemas, structs, etc. although there is no rush as even if it is submitted to the datastore, it is not read-back as we don't use it yet.