CRIMAP-385 Remove dependency on unneeded fixture
This is part of an overall work to rationalise and remove any unneeded, unused, or superfluous fixtures, as these grow a bit out of hand for a while for no good reason.

On Apply I've only detected one superfluous fixture, `application_returned_split_case.json`

It has been replaced by the overarching `application_returned`, with reason variant `split_case`. There is nothing different or complicated that requires having this as a full separate fixture at all.