CRIMAP-387 Support for pruned application struct
This is part of the overall work to simplify the JSON response when listing applications, by removing all unneeded attributes.

As this new representation lacks many of the required attributes, and could even change structure in the future, a new struct has been created to handle it.
Otherwise trying to use the existing struct with the pruned JSON raises multiple errors about required attributes (case_details, provider_details, etc.)

Other than that there is little more to it, as Apply will still find in this pruned struct everything it needs to render the dashboard lists and pagination.