CRIMAP-404 Tweak hearing date validation error
The error message for next court hearing wasn't clear enough, as "today" is a valid hearing date however is not in the future.

Tweaked the error message.
Deleted unused error from date of birth validation.

NOTE: we don't use "The [date of something] ..." for error messages so I've kept it consistent with what we have already.
For example errors for date of birth or offence start/end date etc.