CRIMAP-407 Add postcode label
The postcode lookup step didn't have a specific label for the input, instead it used the h1 of the page as label, but it wasn't entirely clear. Also the hint text didn't make a lot of sense because of this.

As per content/service designers, we are adding a specific label, with text `Postcode`.