CRIMAP-389 Copy changes to existing error pages
As per content designer.

Also changed any references to "back to all applications" to "back to your applications", also as per content designer request.

Added a flash message (that shows on the home page) when the user signs out, this will be the only flash message we will have eventually, but the work to remove the other flash messages and convert to standalone error pages will be done in a separate PR so this one is not too large.

In addition to this, changed the behaviour of the "Start now" button on the home page, so if provider is not signed out, it goes directly to LAA Portal (instead of giving them an unauthorize error like before) and if they are already signed in, direct them to the dashboard.
This also will hopefully reduce the amount of unhelpful 401 alerts we get.