Tentatively fix flaky tests
Although very rare to occur, as we run the tests randomised, there was a scenario where the `FeatureFlags` which is a Singleton was being initialised with an incorrect stubbed `env_name`, and a few subsequent tests making use of some feature flags, were unpredictable.

If this keep happening will investigate further or another solution.