CRIMAP-214 Reset home address
There was an edge case where, if the provider enters a home address, but then for whatever reason they retract from this and want to delete it, there is not easy way to do this.

If they select click the link `My client does not have a home address` the previous entered address would still be there and in the contact details page, they would still see the radio "Same as home address" in the correspondence address options.

This PR improves the situation, but cleaning up any previously entered home address if the user comes back to the postcode step and click the link (which is now a button, styled as a link, so it submits the form) `My client does not have a home address`.

It does not destroy the record as it is needed for the "back" navigation, it only nil all values of the address effectively deleting it.