Relax some prometheus alerts (#390)
Removed one that is not useful at all, if there is any problem with replicas or pods, other alerts will trigger instead.

Removed Medium vulns from Trivy alert, although all the noise so far has been from High vulns, not Medium. But given most vulns are not on our hands (docker/OS/etc.) no point in having Medium alerts right now.

Tweaked the 4xx and 5xx alerts so they are a bit less noisy. Instead of triggering just with 1 error, now it needs to have more than 5 errors in a 5 mins timespan. We can observe how this goes or tweak it until we are happy.