CRIMAP-379 Simplified prometheus exporter approach (#397)
* CRIMAP-379 Simplified prometheus exporter approach

As we are not running Puma in multi process mode (instead we can scale horizontally with more threads and pods), there is a simplified approach to prometheus exporter.

In this version, we don't need a separate process running in a separate pod (meaning no need for having another deployment manifest, or deal with hosts to connect client to server etc.)

The prometheus exporter server instead runs in each pod, in the rails application itself. This have the advantage, we can add custom metrics more easily as we have access the the full rails environment (models, activerecord, etc.) whereas if running the process separately, we don't have it, or we have to load it defeating the purpose.

Added active record and puma instrumentation.

We can see how this approach works and if we consider it is easier, once we implement some custom metrics (next PR). It is relatively easy to change to previous approach.

* Consistent env variable boolean values

* Add metrics port to docker compose