CRIMAP-379 Add applications grafana dashboard
A basic grafana dashboard making use of the recently added custom prometheus metrics.

There is plenty of scope for improvement, this is just an example.

Moved the prometheus rules to the staging namespace so it is easier to iterate them without having to deploy all the way to production (although they can be applied manually anyways).

Added a new prometheus rule to monitor the exporters.