CRIMAP-275 Remove unused DWP check param
Michael has confirmed the `dateOfAward` parameter is no longer needed, in any environment, so it is safe to be removed.

Make sure the `clientReference` can be traced back to one of our applications, so changed to the application UUID (because the applicant record UUID is not persisted anywhere post-submission).

Few other non-functional tweaks.