Redirect to dashboard if user is signed in (#406)
When a use (provider) is not signed in and they access the service through a link or by entering the URL, they see the home page with the start button.

If they are signed in, they get redirected to the applications dashboard instead. This happens if they click the header link (service name) or if they go to the root URL, etc.

This is a more consistent behaviour than showing again the "Start now" button. Simplified the view and removed the partial.