CRIMAP-426 PoC offence list synonyms
The suggestions component we are using for the offences list, comes with a synonym capability, so for instance entering a synonym will return the “expanded“ offence from the list.

We just needed to "create" these synonyms.

This is a functional PoC.

The synonyms are separate and not coupled to our offences list, and work more like a progressive enhancement (because it will only work with the javascript version of the suggestions list anyways).

Everything else will continue the same, the synonym is just a quick way to bring up the “expanded / long“ offence name in the dropdown for providers to select.