Fix cookie banner JS module blocked by browsers
Some browsers come with built in extra privacy measures, and also some plugins can be used to block some "annoyances" like cookie banners.

Main issue was, the blocking was stopping other genuine modules from being properly loaded by the browser, making some functionality that relies on JS fail, like the offences list, or the print button.

Seems to be "fixed" by obfuscating the name of the JS module so it doesn't contain keywords that are likely to be blocked, like `cookie`, `banner` or `consent`.