CRIMAP-427 Validate format of previous MAAT ID
Simple validation to avoid fat fingers or other issues.

The attribute is optional so we still allow blanks. But if the attribute is filled, then we validate is a numeric 6-9 digits.

Ensure all existing regex in other places use `\z` (lowercase) instead of `\Z` (uppercase) as:
> \Z Matches the end of the string unless the string ends with a "\n", in which case it matches just before the "\n".

Which means by using \Z someone could sneak in a newline character and still pass validation.