CRIMAP-429 Refactor error reporting using Rails new API
This is a first part of at least a couple PRs intended to remove some technical debt by refactoring exception capture and error reporting using the modern Rails error reporting API introduced in Rails 7+

https://edgeguides.rubyonrails.org/error_reporting.html

Functionally it should be the same as before, but some boilerplate code is simplified.

Only thing I noticed is the Rails.error.handle, Rails.error.report etc methods do not log the exception, which makes sense, but without it, the exception is quietly hidden and particularly in localhost it can be misleading.

I've made a basic error subscriber just to perform the `Rails.logger.error(e)` so it behaves like before, adding one line of backtrace for additional context.