CRIMAP-430 Swap order of the URN step
Move URN step after the date stamp (i.e. after case type and/or appeal details).

Update task list so the case details task starts with the case type and not with the URN as it was until now.

For consistency, moved also the place where the URN shows in the review application page (and caused a bit of a mess in the tests, sorry).

Everything (except the order, that is) should continue working as before, nothing really has changed except the order of the steps.