Replace cryptic Omniauth error flash alert (#429)
Although it should be very rare and edge case, there could be sometimes Portal authentication errors.

We were showing a rather obscure error message in a flash alert. This has now been replaced with the generic "unhandled" error page, and also we send the exception to Sentry so we are alerted.