Remove `hmcts_common_platform` gem dependency
Based on some recent conversations, and the fact the repo for `hmcts_common_platform` is archived and apparently unmaintained and outdated, one immediate easy thing we can do is bring over the (simplified) CSV file and use the same approach we used for offences.

Functionally everything should continue working as before, and we have more flexibility to update or tweak the CSV file to add/remove or rename courts.

In the future we can plan to move to an API based approach using FaCT.