CRIMAP-439 Ensure enter key triggers main submit action
Default browser behaviour when user focus in a form input, like a text field, and they press enter, is to submit the form.

This usually works just fine, as there is only one submit per form or the main submit comes before the others.

However, in some steps, like the postcode lookup, there is more than 1 submit action and the browser picks the first one in order of appearance in the DOM, which might not be the one that actually performs the “save and continue” (or the postcode search in this case).

This works around this behaviour in a progressive enhancement way with a bit of JS, by cloning the main submit action and positioning it (hidden, even for screen-readers) right at the top of the form so `enter key` picks this one always.

Only forms tagged with "data-multi-action-form=true" will have this progressive enhancement applied, as there is no need to do this with other forms.