Set `Cache-Control` response header to no-store
It avoids possibly exposing personal information when using the browser back button after the current session has been terminated.

This is similar to the approach taken on Review however instead of adding it to all controller actions, this is only set on those exposing possible personal information.

There are many other pages on Apply (all footer pages, and some others) that are not sensitive.

Note pragma header not set as is deprecated and no longer recommended.
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Pragma