Discern between invalid session and invalid token errors
Omniauth authentication could fail due to an invalid authenticity token. For instance, by re-submitting the same form twice.
In this scenario we were redirecting to the unhandled error page, which is not correct and confusing to the user.

Now we just let `InvalidAuthenticityToken` exceptions bubble up, so it can be rescued and handled accordingly within the generic error handling block.

Also previously we were bunching together the invalid session and the authenticity token errors, but they are not really the same and to a user's point of view, showing the invalid session error when the error was produced by an invalid authenticity token (forgery protection) might be confusing as we tell them:

"For your security, we signed you out
This is because you were inactive for 48 hours."

Now each error is separate and the `InvalidAuthenticityToken` will redirect to the more lax and generic "Sign in to continue" page.