CRIMAP-460 Minor tweaks to logging
In the `filter_parameter_logging` we had `details` which is a too broad term that was hindering some logs unnecessarily.

For instance, whole steps like `contact_details`, `address_details`, `appeal_details` or `hearing_details` where being filtered, however there are attributes we don't need to filter in those.

Also city and country are not PII and can be left unfiltered.

Added the signed in provider selected `office_code` to each log trace for better debug purposes.