CRIMAP-462 Datastore health engine routes
Use the datastore health rails engine on Apply to act as a proxy for the health and ping endpoints given the datastore is an internal service only accessible from the cluster.

Moved the `require 'datastore_api'` to the Gemfile declaration as otherwise it wasn't properly loading the engine, I think it may be because the name of the gem is different to the name of the lib require.

Note: the gem ref to be changed with the tag 1.1.0 when new tag is released.