CRIMAP-506 Providers prometheus custom collector
Similar to the existing applications collector, this new collector will expose a few basic metrics about providers as they are being enrolled and use our service.

Added a new panel and widgets to the existing Grafana dashboard so all is in one place.

Note until the collector is deployed, the dashboard will show no data.