Refactor custom prometheus collectors
As we now have 3 collectors, a bit of refactor to DRY some code was due.

Functionally they are exactly the same and produce the same metrics with the same names as before.

Added a new collector to gather the total number of unique offices across all enrolled providers and a Grafana widget to visualise this.