CRIMAP-529 Add MOJ javascript and dropzone
Add all the stuff needed to render an MOJ multi-upload component (dropzone behind the scenes).

Note in order to do this we now have to add to the application the MOJ javascript plus dropzone dependencies.

I'm raising this PR just to have it all set, but is obviously not functional, and I doubt it can be functional out of the box without modifying MOJ javascript because for once, the JS is not propagating the authenticity token and I don't see a way to configure this, meaning only way to propagate the token is to modify the javascript.

My opinion is we shouldn't be using MOJ component, which couples us to their javascript (and requires jQuery nonetheless) and instead code our own JS handling and apply our own styling (which can be the same as the MOJ styling if we wish).