CRIMAP-529 Dropzone without MOJ component
This is a counterpart to previous PR #460.

We may decide this is a better way, as I do believe we will need to heavily modify and customise MOJ multi-upload component anyways, so we might as well avoid using it, and instead just use a standalone dropzone module (that doesn't require jQuery) and our own styling and JS for events handling.

I've made a head start on styling and JS events preparing the structure and foundation. We can either migrate styling and JS handling from MOJ component, or from other service like Civil Apply, or Tax Tribunals, etc. that use dropzone standalone too.

Note how, out of the box, without any additional JS, dropzone handles the Ajax call, and the authenticity token is propagated, etc.