Refactor evidence data model removing bundle abstraction
Bit of a refactor before this becomes too entangled to remove the DocumentBundle intermediate table and abstraction as it seems is of no use and makes code a bit more convoluted and creates additional indirection.

To the best of my efforts everything should continue working as before, no changes to external URLs etc. have been made, and is all internal.

As part of this I've taken the opportunity to introduce in the Documents DB table 2 new attributes:

`submitted_at` - can be populated on rehydration from a submitted application, and this can signalise which documents are already submitted and thus we shouldn't do a hard delete on S3 (just soft delete setting the `deleted_at` attribute)

- `annotations` jsonb attribute that initially will be used for the virus scanner to annotate files that have been scanned and its result. Can be used for some additional functionality later on.