CRIMAP-596 ClamAV in a separate container PoC
Just a PoC to validate assumption we can have ClamAV and signatures in a container while communicating with it through the TCP socket and the `clamdscan` CLI tool.

Using HMPPS docker image because it is already pre-configured and works out of the box but we can use any other.
https://github.com/ministryofjustice/hmpps-utility-container-images/tree/main/hmpps-clamav

Using the most bare bones gem I could find for it, Clamby.

Only "tricky" thing is we have to provide individual config files dependent on environment. The port does not change but the address will, and couldn't find a nice way to do this via ENV variables, as `clamdscan` expects as input a config file.

Not done the kubernetes part yet, until we validate this can work.