CRIMAP-598 First court hearing selection
Follow-up to PR #493 to expose the radio options in the existing hearing details page, for the provider to select whether the first hearing court is the same as the next hearing court and logic to move to next page.

Includes validation as well as logic to delete previous answer if the provider changes the radio selection.

This is now part of the wider journey, but there is not Check Your Answer yet, that will be part of the next PR.
Also missing submission to datastore/rehydration.

This is not behind a feature flag as I'm hoping to finish it soonish enough to be able to test and release as is a minor iteration on current form.