CRIMAP-646 Refactor deletion of evidence documents (#496)
* CRIMAP-646 Refactor deletion of evidence documents

This is a first pass on fixing/adapting the delete functionality to the latest agreed behaviour.

Evidence can be hard-deleted while it is still in draft (never submitted) mode. Once an application is submitted, its evidence can still be deleted in a resubmission, but is a soft-delete, meaning it remains in the S3 bucket as previous version(s) of the application will still reference these files.

In order to do this on rehydration we just set the `submitted_at` attribute in the Document object, as a way to signalise this was a previously submitted document.

Refactor the `Delete` service object to protect against accidental (hard)deletions of submitted files, and also to not try to delete files that have not been stored really (`s3_object_key` is nil).
Better handling of errors.

Extracted the application purge functionality to a little service object that now also performs cleanup of S3 documents if necessary (and according to rules above) before destroying the local application record along with all their relationships.

* Protect documents about to be submitted

Make sure we are not deleting stored documents that are about to be submitted!

* Do not use purger in the submission

This simplifies the cognitive load of having a purger that does different things in 2 separate places.

Essentially keep the submission as was before, nothing wrong with it, and the purger will be used only for drafts (deletions through the dashboard).