CRIMAP-644 Show benefit type in CYA page
Added the benefit type to the check your answer page, in place of the previous "Passporting benefit" essentially merging both rows into one as confirmed by content designer.

If the question is not asked (under 18s) then the row will not show up.

Ensured `benefit_type` attribute is reset in a couple places if some key details are changed, as this value is linked to the DWP benefit checker.

Feature is still being feature flag, need to be removed if we want this in production at some point in the future once QA'ed.