CRIMAP-600 Add first court hearing details to CYA
Follow-up to previous PRs #493 and #494 to add the first court hearing details (if any) to the check your answers page.

Also bump to the schemas version where this new attributes are part of the Struct to be able to read back submitted applications.

If the first court hearing question was not asked, we hide the row and section.

It is important to merge this datastore PR too otherwise submitted applications will not have the new attributes:
https://github.com/ministryofjustice/laa-criminal-applications-datastore/pull/175