Remove `benefit_type` step feature flag
In the event we want to release this.

As part of this PR I've also re-wired the logic for retrying the DWP check when it fails due to connectivity or unhandled exceptions, so it does not go again to the benefit type page, and instead just re-try the DWP check as intended.