Reconcile latest changes in main branch
Removed for now the `expected_evidence` attribute, as it may be early to know how to really use this. I'm trying to avoid a situation similar to the passporters where the logic might sometimes not be run and we end up with out of sync values.

Perhaps a better approach is to always calculate the expected evidence based on the answers provided, instead of keeping a persisted list.

The list can then be calculated from the answers. The answers are needed regardless because we need to rehydrate returned applications, thus why I've persisted `has_benefit_evidence` answer.