CRIMAP-576 Add `has_benefit_evidence` page (#492)
* Add has_benefit_evidence_page

If benefit check fails, details are correct, user can register that
they have evidence they intend to upload later
Adds virtual attribute, only needed locally in apply to resurface later
at evidence upload step.

* Reconcile latest changes in main branch

Removed for now the `expected_evidence` attribute, as it may be early to know how to really use this. I'm trying to avoid a situation similar to the passporters where the logic might sometimes not be run and we end up with out of sync values.

Perhaps a better approach is to always calculate the expected evidence based on the answers provided, instead of keeping a persisted list.

The list can then be calculated from the answers. The answers are needed regardless because we need to rehydrate returned applications, thus why I've persisted `has_benefit_evidence` answer.

---------

Co-authored-by: Jesus Laiz <zheileman@users.noreply.github.com>