CRIMAPP-1499: Remove back link from completed application page
Now that search requests use POST, it's better for users to rely on
the browser's native back button to return to the search results page.
This ensures the search form remains populated.

The GOV.UK Design System advises using back links for question pages
and other pages within a multi-page transaction when appropriate.
However, it does not apply in this case.