CRIMAPP-1712: Use dedicated readiness endpoint in staging
Previously, the health endpoint was used for both readiness
and liveness checks. #1395 added clamav check to the health endpoint.
This commit reverts that change to the health endpoint and adds a
dedicated readyz endpoint to determine if clamav is available to the web app.
This allows production continues to function as it did before #1395 and
also prevents the staging service from being taken down due to transient
issues with clamav after deployment.